# assiduidade_system_backend/openapi.yaml
openapi: 3.0.3
info:
  title: FaceClock API
  version: 1.0.0
  description: API principal do MVP de registro de ponto com reconhecimento facial.
servers:
  - url: https://api.faceclock.local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
...
    ClockRegisterRequest:
      type: object
      required: [idempotency_key, user_id, device_id, event_type, recorded_at, source]
      properties:
        idempotency_key:
          type: string
        user_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        recorded_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [ONLINE, OFFLINE_SYNC, MANUAL, INTEGRATION]
        confidence_score:
          type: number
          format: double
        liveness_score:
          type: number
          format: double
        geo_lat:
          type: number
          format: double
        geo_lng:
          type: number
          format: double
    ClockRecord:
      type: object
      required: [id, user_id, event_type, recorded_at, source, sync_status]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        recorded_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [ONLINE, OFFLINE_SYNC, MANUAL, INTEGRATION]
        sync_status:
          type: string
          enum: [SYNCED, PENDING, FAILED]
        confidence_score:
          type: number
          format: double
        liveness_score:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
    ClockSyncRequest:
      type: object
      required: [record_ids]
      properties:
        record_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    ClockSyncResponse:
      type: object
      required: [synced_record_ids, not_found_record_ids, total_requested, total_synced]
      properties:
        synced_record_ids:
          type: array
          items:
            type: string
            format: uuid
        not_found_record_ids:
          type: array
          items:
            type: string
            format: uuid
        total_requested:
          type: integer
        total_synced:
          type: integer
    AdjustmentRequestInput:
      type: object
      required: [reason]
      properties:
        clock_record_id:
          type: string
          format: uuid
        requested_event_type:
          type: string
          enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        requested_recorded_at:
          type: string
          format: date-time
        reason:
          type: string
...
    AdjustmentRequest:
      type: object
      required: [id, user_id, status, reason, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        clock_record_id:
          type: string
          format: uuid
        requested_event_type:
          type: string
          enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        requested_recorded_at:
          type: string
          format: date-time
        reason:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED]
        review_notes:
          type: string
        reviewer_id:
          type: string
          format: uuid
        reviewed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    PaginatedClockRecords:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ClockRecord'
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
...
paths:
...
  /api/v1/clock/register:
    post:
      tags: [Clock]
      security:
        - bearerAuth: []
      summary: Registra evento de ponto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockRegisterRequest'
      responses:
        '201':
          description: Registro persistido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClockRecord'
        '409':
          description: Chave idempotente ja utilizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/clock/sync:
    post:
      tags: [Clock]
      security:
        - bearerAuth: []
      summary: Sincroniza registros de ponto pendentes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClockSyncRequest'
      responses:
        '200':
          description: Resultado da sincronizacao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClockSyncResponse'
  /api/v1/clock/me:
    get:
      tags: [Clock]
      security:
        - bearerAuth: []
      summary: Lista historico do colaborador autenticado
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Historico retornado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedClockRecords'
  /api/v1/clock/adjustments:
    post:
      tags: [Clock]
      security:
        - bearerAuth: []
      summary: Solicita ajuste de ponto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustmentRequestInput'
      responses:
        '201':
          description: Ajuste criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdjustmentRequest'
        '403':
          description: Tentativa de ajustar registro de outro usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Registro de ponto nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados insuficientes para criar ajuste
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...